Significantly reduce the timeout app_jit is given on the buildbot
Currently tests in the app_jit-debug configuration are given 16 minutes
timeout. This caused us to run into a global timeout of test.dart (it's
"debug timer" which is set to 10 minutes of inactivity).

Even the slowest tests run in the app_jit-debug configuration within around 1:30 m on the buildbot.

This CL changes the timeout from 16 minutes to 4 minutes.

R=vegorov@google.com

Review-Url: https://codereview.chromium.org/2685303002 .