[dart:io] flush() stdin before close()ing
Calls to close() on an IOSink do not imply that writes buffered by the
underlying StreamConsumer have been flushed. This CL adds a note
indicating this to the IOSink docs, and adds calls to flush() before
calls to close() for some calls to stdin.close() for spawned
processes in our tests.

related #28737

R=fschneider@google.com

Review-Url: https://codereview.chromium.org/2687963003 .