Add the ability to convert an analyzer token stream to a Fasta one.
This will be needed in order to integrate Fasta into the existing
analyzer codebase, because the analyzer task model considers the
parser and scanner to be separate black boxes with an analyzer token
stream in between.  In order for Fasta to be able to drop into those
black boxes, we will need to be able to convert Fasta tokens to
analyzer tokens after scanning, and then back to Fasta tokens prior to
parsing.

Tested via a "round trip" test that converts the analyzer token stream
into a Fasta token stream and then back.

Note that the transformation is not perfect; it currently fails to
recreate error tokens and it fails to hook up the connection between a
BeginGroup token and the corresponding end token.  But it is enough to
be able to pass tests for now.  It will be improved (and tests will be
expanded) in future CLs.

R=ahe@google.com

Review-Url: https://codereview.chromium.org/2689733003 .