Create a new class for parser tests that test top level constructs.
These parser tests should be easy to port over to Fasta; the others
will be more difficult because they invoke parser methods that may not
have exact analogues in Fasta's parser.

This CL is almost entirely code motion; the only substantive changes
are the creation of the new class TopLevelParserTest, and the creation
of the new mixin ParserTestHelpers (for the helper methods needed by
TopLevelParserTest).

R=brianwilkerson@google.com, scheglov@google.com

Review-Url: https://codereview.chromium.org/2703553003 .