cleanup --batch-mode sourceFiles
BuildMode, the dartanalyzer driver used when the "--build-mode" option
is supplied, expects sourceFiles in the format "<uri>|<filepath>",
but the rest of the dartanalyzer code base does not understand this format.
This change updates BuildMode to rewrite sourceFiles,
stripping the "<uri>|" prefix, so that it does not cause problems
with code that does not expect this format.

Fix https://github.com/dart-lang/sdk/issues/28889

R=brianwilkerson@google.com

Review-Url: https://codereview.chromium.org/2720493002 .