Adds tools/ninja.py that runs goma builds in parallel
This CL adds a script tools/ninja.py that runs Goma enabled builds
in parallel. Non-goma builds are run serially, as before.

Conflating this feature with gyp support in tools/build.py
got to be a bit of a mess, which is why this is split off as a
separate script. This script can replace build.py after gyp support
is removed.

With a warm goma:
$ time ./tools/build.py -m debug,release -a x64,ia32 runtime runtime_precompiled
...
real	4m25.627s
user	3m30.740s
sys	1m3.873s

$ ./tools/ninja.py -m debug,release -a x64,ia32 runtime runtime_precompiled
...
The build took 108.900 seconds

R=rmacnak@google.com

Review-Url: https://codereview.chromium.org/2712743009 .