Create a TokenStreamRewriter class.
This class will allow parser error recovery to insert synthetic tokens
into the token stream.  It makes use of Token.previousToken when it's
available, but does not depend on it; when it's not available, it
falls back on heuristics which should still be fast, and which don't
impose any performance burden until the first synthetic token is
inserted.

R=danrubel@google.com

Review-Url: https://codereview.chromium.org/2732193007 .