Change the calling conventions for handleUnrecoverableError.
Instead of returning the token from which parsing should proceed, it
returns a token whose "next" pointer points to the token from which
parsing should proceed.

The rationale is that when error recovery is performed in the context
of the analyzer, the listener will need to insert synthetic tokens
into the token stream which will then be passed to listener methods;
the change in calling conventions allows the listener to return the
synthetic token to the parser.

The other listeners don't need to insert synthetic tokens into the
token stream; they just need to create a synthetic token whose "next"
pointer points to the token from which parsing should proceed.  The
convenience method "newSyntheticToken" is added for the benefit of
these listeners.

No code is affected outside of error handling, so the performance
impact when compiling correct code should be zero.

R=ahe@google.com, scheglov@google.com

Review-Url: https://codereview.chromium.org/2738903002 .