Add fasta-based summary generation code.
The code is still a work in progress, so its test is marked as failing.

Note: I believe there is code duplicatoin between the StackListener
class introduced by this CL and the StackListener class in
pkg/front_end.  Once more tests are passing, I intend to remove this
code duplication.

R=sigmund@google.com

Review-Url: https://codereview.chromium.org/2738273002 .