[kernel] set dartAsyncMarker to Sync for continuations
After 89b68e905b6a096148cedc580016a595e106dcdd debugging of async
functions via kernel haven't worked.
Prior to this the dartAsyncMarker for e.g. a rewritte async function
would have been Sync, after it was Yielding.
This changes it back in the continuations rewriting cases.
An alternative fix would be to change
  function.set_is_debuggable(
      kernel_procedure->function()->dart_async_marker() == FunctionNode::kSync);

to something like
  function.set_is_debuggable(
      kernel_procedure->function()->dart_async_marker() == FunctionNode::kSync ||
      kernel_procedure->function()->dart_async_marker() == FunctionNode::kYielding);

in the C++ code.

BUG=
R=kmillikin@google.com

Review-Url: https://codereview.chromium.org/2743283002 .