Script finding the build bot group for a given shard.
Helping to solve a silly little problem that I've experienced many
times: If you learn that there is a problem with a buildbot shard like
'dart2js-linux-drt-2-2-be' it may not be obvious that it should be
found on the buildbot status page (go/darto) in the group 'chrome'.
I didn't really know any other way than hovering the mouse over a few
dozen areas on that web page, hoping that one of them showed that
shard name as a tool tip.

This CL adds a small script, 'bin/shard2group.dart' to the 'gardening'
package, which will find the group from any given (partial) shard
name. For instance:

  > dart bin/shard2group.dart noopt                                                                                                                                                                                                                                                [16:00]
  vm-precomp(1): vm-noopt-simarm64-mac-be
  vm-precomp-dev(1): vm-noopt-simarm64-mac-dev
  vm-precomp-stable(1): vm-noopt-simarm64-mac-stable

The number indicates the placement of this shard in the group.

The script 'bin/create_shard_groups.dart' transforms the source code
for the buildbot web page to the source code for a Dart part file
which provides 'bin/shard2group.dart' with the concrete data on the
build bots and their shards.

R=johnniwinther@google.com

Review-Url: https://codereview.chromium.org/2756193002 .