Record type inference errors into linked summaries.
In all cases, if initializer inference fails, it is an error and produces
no result. Implementations are free to treat this as a result of dynamic
after emitting the error for the purposes of emitting subsequent error
messages, but the behavior of programs for which initializer inference
fails is unspecified.

R=brianwilkerson@google.com
BUG=

Review-Url: https://codereview.chromium.org/2763673002 .