Introduce a shadow hierarchy and factory class for use by fasta's BodyBuilder.
In the long run, this shadow hierarchy is intended to be used both for
kernel and analyzer ASTS; this should allow a single BodyBuilder class
to build either a kernel or an analyzer AST depending on which factory
it's connected to.

For now we only have a kernel shadow hierarchy, and it's incomplete,
so the code is not yet used.  I intend to follow up with CLs that
complete the implementation and hook it up to BodyBuilder.

R=ahe@google.com, scheglov@google.com

Review-Url: https://codereview.chromium.org/2769723004 .