Use canonical names for some name-based lookup.
For compiling the platform libraries separately and for hot reload, we
will not link canonical names of Kernel libraries, classes, and
members to their Kernel definitions.  Those definitions will not be
persistent.

As a first step, change name-based lookup of libraries and classes to
be be based on the canonical name instead of the Kernel definition.

BUG=
R=jensj@google.com, kustermann@google.com

Committed: https://github.com/dart-lang/sdk/commit/a15b8c49ea195345f542d4efbe8d6dd29268d030
Review-Url: https://codereview.chromium.org/2762213002 .