Fix for Quick Fix to import SDK classes.
It's not 100% replacement, we handle SDK libraries just as any other
library, if it has not been imported yet somewhere, we don't know
about it. But probably this covers the most often case of using a
library in a new file of the project that already uses this SDK
library.

R=brianwilkerson@google.com
BUG=

Review-Url: https://codereview.chromium.org/2781963002 .