VM: Fix incorrect canonicalization rule for `UnboxedIntConverterInstr`.
The rule assumed that `UnboxedIntConverter(U->V, UnboxedIntConverter(V->U, v))` is
equal to `v` which is incorrect if the first convertion is truncating and
V is a larger representation (Mint) than U (Int32 or Uint32).

Now canonicalization rule takes into account if the first convertion is
truncating and does not try to short-circuit such convertions.

Fixes https://github.com/dart-lang/sdk/issues/29137

R=kustermann@google.com
BUG=http://dartbug.com/29137

Review-Url: https://codereview.chromium.org/2793003002 .