Clean up documentation of Platform in dart:io.
The class is just a static-only container, and should really be rewritten as
either top-level getters instead of static getters, or as instance getters
on an interface (in case we ever want to mock it).

Also, the isOSName getters for each individual supported OS doesn't scale.
We should consider removing them and only support the `operatingSystem` getter.

R=floitsch@google.com

Review-Url: https://codereview.chromium.org/2751913006 .