VM: [Kernel] Implement CatchBlockEntryInstr::should_restore_closure_context on all platforms.
This was the only Kernel specific backend change - to simplify the graph builder
we allowed it to capture exception and stack trace variables. This in turn
requires CatchBlockEntryInstr to respect that this variables can be captured
and correctly update them when exception is caught.

This was implemented on X64 and ARM before. This CL expands implementation to
all other implementations including DBC.

R=zra@google.com

Review-Url: https://codereview.chromium.org/2796283003 .