Don't visit a frame's code object twice.
One slot in each Dart frame contains a "pc marker". Prior to the calling convention change, the pc marker was a fixed offset from the code's entry point, from which the code object could be derived. After the change, the pc marker is a pointer to the code object itself.

R=asiva@google.com

Review-Url: https://codereview.chromium.org/2804793003 .