Temp fix to allow plugins to serve the getErrors APIs.
Had to allow a means of supplementing existing errors/line info, but
then also for where dart has no response whatsoever (like html), need a
way to provide analysis options and line info in addition to errors.

Happy to clean it up, but also not very motivated to make it perfect
since it should be temporary.

BUG=
R=brianwilkerson@google.com

Review-Url: https://codereview.chromium.org/2815643004 .