Flush on write in FileByteStore.
We see (rarely, but still) situations when FileState attempts to read
a *.unlinked file, but the file is empty. We think that this happens
when the user's machine crashes, so files are created, but not fully
written to disk.

This change has its cost - increase of FileByteStore.put() CPU usage
from 7% to 10% on Mac.

R=paulberry@google.com, brianwilkerson@google.com
BUG=

Review-Url: https://codereview.chromium.org/2821683002 .