Make copy of the drivers list in SearchEngineImpl2 to avoid ConcurrentModificationError.
I think this happens when a driver is disposed why search is in
progress. So, we need to make defensive copies.

This also makes me realize that we don't handle situations when
a driver instance is disposed, but it has pending requests, such as
getResult(), getIndex(), etc - actually any getXXX(), and search. As
it is now, I think we just leave these Future(s) and they never
complete.

R=brianwilkerson@google.com
BUG=

Review-Url: https://codereview.chromium.org/2818163003 .