Remove Compiler/JavaScriptBackend from metadata_collector
- and Compiler from type_test_registry.

The CodeEmitterTask was removed from the TypeRepresentationGenerator to
break a dependency cycle between the two. An Emitter is now provided as
an argument to TypeRepresentationGenerator methods instead.

This is a step towards passing data object to the codegen/emitter
instead of pulling data directly from Compiler/JavaScriptBackend. This
is needed to support an element model switch between codegen and
resolution.

R=efortuna@google.com

Review-Url: https://codereview.chromium.org/2809203003 .