Prepare to publish analyzer, front_end, and kernel.
Due to the tight coupling between analyzer and front_end, they need to
be published atomically.  The last published version of kernel is old
enough that it needs to be published too.

I've bumped the versions to the following, and changed the
dependencies so that this set of versions is mutually compatible:

- analyzer: 0.30.0-alpha.3
- front_end: 0.1.0-alpha.2
- kernel: 0.2.0

(Note that kernel's version didn't need bumping since the most
recently published version of it is 0.1.0)

R=asgerf@google.com

Review-Url: https://codereview.chromium.org/2828273003 .