Reland "Use off-heap data for type feedback in PolymorphicInstanceCallInstr"
This reverts commit 2135c8a90f37091536bb6707edc5bf5a48bf315c.
This is a reland of https://codereview.chromium.org/2809583002/

Original text:
We have been using IC data for receiver type info in the
PolymorphicInstanceCallInstr. This is a data structure optimized for access
from hand-coded assembly stubs, and placed on the GC-ed heap, which means it
has to be accessed through handles. With this change we move it to the zone
memory, which means it can be freed without an old-gen GC.  As a side-effect
the zone arrays use exponential growth for amortized constant space allocation
instead of growing by 1 for quadratic allocation when we add classes, further
reducing memory pressure.

R=vegorov@google.com
BUG=

Review-Url: https://codereview.chromium.org/2833333003 .