Add AnalysisDriver.getLibraryByUri().
We need this for DDC, and it actually uses it only for SDK - both in
cases when DDC compiles SDK (so it needs to analyze SDK libraries),
and in when DDC compiles something other than SDK (so it resynthesizes
SDK classes from the provided external summary store). In all cases
DDC needs these SDK libraries to build TypeProvider and do some HTML
magic.

R=brianwilkerson@google.com
BUG=

Review-Url: https://codereview.chromium.org/2835123004 .