Move dartk to the tool folder
"bin" is part of the public API of a package, private tools can go under the
`tool` folder. This allows us to use a dev_dependency on analyzer without
exposing a broken import publicly.

I've only moved `dartk` so far (which I expect will be removed once we have a
better API for front_end.

R=asgerf@google.com, kmillikin@google.com

Review-Url: https://codereview.chromium.org/2842643002 .