[Gardening] Increase timeout for front_end_inference_test
This test is around 30 seconds on the built bot and hits sometimes
package:test/test.dart's 30 second timeout.

Unfortunately the test deviates significantly from normal tests due to
the use of a custom dart:mirror-based test framework, it's hard to use
the normal mechanisms for increasing the timeout (e.g. using @Timeout
annotation, or specifying the timeout on the 'test(name, () {})' call)

  -> We just wrap it in a group with a different timeout.

Review-Url: https://codereview.chromium.org/2848233002 .